Authentication failures fixed on Auctions Service.

+++ src/Services/AuctionService/Application/Auctions/Delete.cs
@@ -10,7 +10,8 @@ public class Delete
 {
     public class Command : IRequest<Result<Unit>?>
     {
-        public Guid Id { get; set; }
+        public required Guid Id { get; set; }
+        public required string User { get; set; }
     }
 
     public class Handler : IRequestHandler<Command, Result<Unit>?>
@@ -26,11 +27,19 @@ public class Delete
 
         public async Task<Result<Unit>?> Handle(Command request, CancellationToken cancellationToken)
         {
+            var auction = await _auctionsRepository.DetailsAuction(request.Id, cancellationToken);
+            if (auction == null) return null;
+            if (auction.Seller != request.User)
+            {
+                var result = Result<Unit>.Failure("Forbid!");
+                result.ErrorCode = "403";
+                return result;
+            }
             _auctionsRepository.DeleteAuction(request.Id, cancellationToken);
             await _auctionsPublisher.PublishAuctionDeleted(request.Id);
 
-            var result = await _auctionsRepository.SaveChangesAsync(cancellationToken) > 0;
-            if (!result) return Result<Unit>.Failure("Failed to delete the auction!");
+            var saveChangesResult = await _auctionsRepository.SaveChangesAsync(cancellationToken) > 0;
+            if (!saveChangesResult) return Result<Unit>.Failure("Failed to delete the auction!");
             return Result<Unit>.Success(Unit.Value);
         }
     }
